requirements.txt
aiofiles==23.2.1
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
build==1.2.2.post1
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
decouple==0.0.7
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.7
django-extensions==4.1
django-grappelli==4.0.2
django-jazzmin==3.0.1
django-pwa==2.0.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
ffmpeg==1.4
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
fsspec==2025.3.2
futures==3.0.5
gradio_client==1.8.0
gunicorn==23.0.0
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
hyperframe==5.2.0
id==1.5.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mechanize==0.4.10
oauthlib==3.2.2
oktoncss==0.1.1
packaging==24.2
platformdirs==4.3.7
polib==1.2.0
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.9.0
pyproject_hooks==1.2.0
python-decouple==3.8
python-dotenv==1.0.1
python-for-android==2024.1.21
python-pip==1.1.1
python-telegram-bot==22.1
python3-openid==3.2.0
pytube==15.0.0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==14.0.0
runware==0.4.8
setuptools==58.5.3
sh==1.14.3
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.4.0
soupsieve==2.6
sqlparse==0.5.3
stripe==12.1.0
toml==0.10.2
tqdm==4.67.1
twine==6.1.0
typing_extensions==4.13.1
tzdata==2025.1
urllib3==2.3.0
virtualenv==20.30.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
yt-dlp==2025.3.31
